ciao-controller: datastore: do not use defer to unlock mutex in AddPool()
In the error path for AddPool(), we need to delete the added pool
if there was a failure. This function does it's own locking of the
pools mutex, so in order to ensure that we don't deadlock, we need
to first unlock the mutex prior to calling. Because of that, we cannot
use a defer on the Unlock() call. Remove the defer statement, and
add the explicit Unlock() call to all the error paths in the AddPool()
function.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>