fix comments on ssntp error versus event
A series of SSNTP Error type frames are mis-identified in comments in
the ciao-scheduler and testutil ssntp servers.  They are referred to as
"event", which is a specific ssntp frame type distinct from "error".  This
mis-identification can lead to confusion regarding what must be explicitly
forwarded by an ssntp server versus is automatically propagated.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>