ciao-cli: multi-subarg-parsing
To extend ciao-cli to allow multiple "-volume" arguments on during a
"ciao-cli instance add" call, I need a semi-complex usage of the flag
package.  This commit comes with a massive thanks to Jim Lawless' blog post at:
https://lawlessguy.wordpress.com/2013/07/23/filling-a-slice-using-command-line-flags-in-go-golang/
without which I could not cmoprehend using the golang flag package on a
flag.Var slice of subarguments values.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>