ciao-controller: allow overlapping subnets for tenant internal IPs
We no longer need to prevent overlapping subnets for internal IP
addresses between tenants. Modify the algorithm to allocate
internal IPs to allow tenants to have overlapping/duplicate IPs.
Remove the test for overlapping IPs since we don't need it anymore.

We had previously disabled overlapping IP's due to a bug in docker
which was not sending information to the plugin correctly. Now that
docker has fixed this bug, we can re-enable this feature.

Launcher requests docker to create & launch the container, fully
specifying the networking configuration.

docker run -it --net=<subnet.Name> --ip=<instance.IP> --mac-address=<instance.MacAddresss> ...

There was a bug in docker 1.10.03 (which has been fixed in the 1.11
dev version) which does not pass the --ip parameter to the remote
IPAM plugin.

Fixes: #712

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>