ciao-scheduler: glog before init leaves logs in /tmp
The scheduler was initializing its log late, which meant glog tried
to write to a non-existant directory and apparently falls back to its
default /tmp target directory.  To fix this I separate the log init helper
code into a helper function and call that right after flag.Parse() and
insure any prints prior to completion of this are straight outputs not
through the glog package.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>