ciao-controller: track user tag in volume description
The block device mapping v2 API allows specifying a tag along with a
volume.  We already have a Description field internally.  When a tag flows
in for a new volume, it needs to end up in that field.  Otherwise it's not
stored and later in a volume list the Description field is not populated.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>