ciao-down: Fix an issue with GOPATH mounting
ciao-down was not updating the /etc/fstab file correctly as there was
an issue with the mounts section of the cloud-init file.  The final
two mount options were not enclosed in quotes.

In addition, it doesn't always mount the hostgo device correctly on the
first boot.  For this reason we include an additional mount command in
the runcmd section, to ensure the GOPATH is mounted.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>