ciao-controller: change Persistent to Ephemeral
Internal to ciao-controller there is a type StorageResource with a
Persistent flag.  Throughout the rest of the codebase and APIs we have a
concept of ephemerality.  This inconsistency leads to confusion which makes
it easy to create a bug which has the flag inverted in some place.  Such
bugs could lead to user data loss.  Making consistent use of a boolean
tracking ephemerality (inverted relative to persistence) means less
likelihood of such bugs and thus less likelihood of accidental data loss
from a bug.

Note: this patch appears to have an extra inversion of ephemeral=true/false
versus persistent=true/false.  That is there to keep a test case passing as
of this patch, but does mark a bug which is resolved in the next commit.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>