Single VM: Remove image service race condition
There was a race condition in the Single VM setup.sh in which it started
controller and that waited for 5 seconds before trying to connect to
the image service.  Sometimes, 5 seconds is not enough, particularly
if you have just set up a new VM with ciao-down.  This commit fixes the
issue by replacing the sleep with a loop that exits when a successful
attempt has been made to contact the image service or we time out.  The
structure of the loop was copied from the code that integrated the
keystone container into setup.sh.

Fixes #922

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>