ciao-controller: honor user specified volume ephemerality/bootability
Prior to me splitting addBlockDevice() out of getStorage, there were a set
of hard coded true/false settings for ephemerality/bootability of a volume
and I just kept those.  But then I noticed that controller_test.go had an
unexpected true/false check and Rob figured it was a copy/paste error in
his commit 37d23ea08bd5ab15ce6f9da16684284e2f14527f which didn't really
matter because the flags were mostly ignored prior to my patches.

What we really want is the user specified config to be honored.  This patch
removes the hard coded sets of s.Ephemeral and s.Bootable, and has the test
check that the ephemerality after getStorage() is the same as what was
input.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>