ciao-down: Check nested KVM is enabled
When running inside a Single VM, Single VM currently requires the host to
support nested virtualization.  Without this feature ciao will be unable
to launch any CNCI instances and so the whole cluster will be unusable.

This commit checks to see if nested virtualization is enabled.  If it is
not, ciao-down will refuse to start or prepare any new VMs.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>