ciao-down: Generate SSH keys instead of using id_rsa
Previously, ciao-down relied on the user having a pre-generated SSH key
called ~/.ssh/id_rsa.  If the key had not been pre-generated by the user
ciao-down would not work.  This commit modifies ciao-down to generate
a new SSH key pair specifically for ciao-down.  It also optimises the
code in certain places.  The public SSH key is now only read when
running prepare.  It is no longer read when executing the other commands.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>