ciao-image: Fix error handling on not-found images
When an image is not found, we need to return a "Not Found" error,
otherwise, we should send the proper error, for example a database
error, it should be handled differently, it means that it would
return a Server Error (500) to the final user becauser it was a
server side error.

The remaining errors 400, 401, 403, 409, 410, 413, 415, and 503
will be addressed once tenants support is addressed in ciao-image
service.

Fixes #566

Signed-off-by: Obed N Munoz <obed.n.munoz@intel.com>