networking: libsnnet: Fix glog dependency relying on clogger
Because a vendored version of glog causes some issues when libsnnet package
is used from external package such as virtcontainers, we introduce clogger
abstraction layer as the logger to use instead of glog.

The external package using libsnnet will have to define its own implementation
of clogger, or it will result in no traces printed. This is due to the default
implementation of clogger used by libsnnet, which is CiaoNullLogger.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>