networking: libsnnet: Set Vnic peer address explicitely
The API SetHardwareAddr() was implicitely creating a Vnic peer related
to the current Vnic, and it was setting the hardware address for this peer.
It was a very specific case for Ciao but we want more explicit things now
that libsnnet is used externally from other projects.

This patch simplifies SetHardwareAddr() function and adds the complexity
in the Ciao's calling code instead.

Also, it returns a default peer name in case the link name does not fit the
expected prefixes. It is needed because in case you set manually the link
name, you will get a creation error. Indeed, the peer name will be "", and
it will be impossible to create a link with that name.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>