ciao-controller: BAT expects that workload storage is ephemeral
Commit c389aed9b22c360a522630a6d76ca71c8424acda breaks the BAT.  This
fixes it.

Previously the workload storage csv had a persistent flag and it was set
to true.  My storage volumes extensions PR changed that to an ephemeral
flag set to false.  But prior to my PR, the user indicated flag was
actually ignored in getStorage().  I introduced code to honor it instead
of ignore and force in some cases persistence and in some ephemerality.
BAT broke because it expected the example workload volumes to be
ephemeral, despite not being flagged as such.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>