openstack: retrieve tenant from Token when absent from mux
when validating token, tenant is obtained from mux.Vars of the
request, on certain cases (such as the image service) the tenant
variable it's absent due to not being specified in the endpoint URI,
this makes ciao to believe we're requesting operation as an admin when
that's not true.

This commit tries to get tenant from token if no tenant is obtained from
mux.Vars and updates the logic validation of the token; checking if the
user from the given token is admin and validating the services the token
can access. If user isn't admin but services are valid, we assume that
the request it's from a simple (non-admin) user.

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>