openstack: requests with invalid admins mean non-admin user's request
Previously, when a request with missing tenant was performed, the user
was assumed to be an admin, this logic caused the image service to fail
by how the procedure to obtain the tenant was done.

Previous commit changes that logic, obtaining a tenant ID when none it's
provided, with that change now request where an invalid admin is given,
are mean to be performed as a non-admin user.

This commit changes a test case in
`openstack/identity/identity_test.go: TestHandler`, where previously
a combination of validServices and invalidAdmins meant an `StatusUnauthorized`
now mean the operation is being performed as a non-admin user instead.

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>