ciao-launcher: Only unmap explicitly mapped volumes and improve logging
The cfg.Volumes slice contains all the volumes used by the instance -
both those that have been attached after boot and those specified on the
qemu command line.

The volumes specified on the command line will be unmapped by qemu when
it is terminated and so those do not need to be explicitly unmapped.

This patch stores a boolean in the volumeConfig struct indicating if the
volume should be unmapped.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>