ciao-launcher: Wait for virtualizer to terminate
After requesting that the virtualizer (e.g. qemu) to stop wait on the
channel monitoring the domain socket used to communicate with the
virtualizer.  This allows us to identify when the virtualizer has
finished and the next step of the cleanup can begin.

This addresses #1009 by not trying to unmap volumes that are still in
use by the running qemu process.

Fixes: #1009

Signed-off-by: Rob Bradford <robert.bradford@intel.com>