ciao-launcher: Fix ciao-launcher unit tests
A small change was needed to the ciao-launcher unit tests following the
fix to #1009.  The instance go routine now expects and waits for the
monitorClosedCh channel to be closed after it has deleted an instance.
If the channel isn't closed the instance go routine blocks for a few
seconds and this blocking caused a timeout in the unit tests.  The unit tests
now close the monitorClosedCh channel ensuring that the instance
go routine does not block after deleting an instance.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>