Networking: cnci-agent: Persist network state across reboots/migration
Add support to the CNCI agent to save the state of the CNCI
network setup to the boot disk. The network state is recovered
by replaying the commands. This allows for potential migration
to a new implemenation of the command if the cnci agent is
upgraded.

This allows the CNCI agent to recover its state if the CNCI
is rebooted or migrated to a different network node.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>