ciao-controller: Add quota service implementation
The quota service provides a mechanism to Consume/Release quotas as
specified by a series of payloads.RequestedResource structures.

Internally it tracks the usage of the associated resources and if the
requested consumption exceeds a previously specified quota then the
service will give feedback that the request should be denied.

It also provides an API for setting quota and getting and usage
information in bulk for the HTTP API.

The service can be safely accessed from multiple goroutines and offers
Consume() feedback asynchronously so as to not give delays when e.g.
starting instances.

Fixes: #1073

Signed-off-by: Rob Bradford <robert.bradford@intel.com>