ciao-controller: Add database code for quotas
This change adds the database code and testing required to update the
quotas and limits in the database and a very basic wrapping of the
database functionality via the datastore.

There is no caching of the quotas details in the datastore as that is
handled by the quota service itself.

Fixes: #1056

Signed-off-by: Rob Bradford <robert.bradford@intel.com>