ciao-down: Add --runcmd option
--runcmd allows users to specify a file that contains a runcmd: section
of a cloud-init file.  The commands in this user specified file get added
to the runcmd: section of the cloud-init file built into ciao-down.  This
allows users to customize their VMs.  Currently, the syntax of this file
is a little obtuse and undocumented.  Going forward I'll make it easier
to write these sort of files and allow users to replace the entire
cloud-init file.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>