ciao-cli: identity: return wrapped errors
The ciao-cli's identity code that handles tokens is never returning an
error.  These leads to strange runtime behaviors in the face of a bad
identity service:  A token related command would fail, but the function
would return a nil error.  The caller would then carry on with
uninitialized "" string variables, which of course led to more failures.
This obfuscates the root issue.

There is an errorf (in some cases of questionable quality) present into
the code adjacent to the nil return.  For simplicity, I've just moved
these existing messages into the errors.Wrap() in addition to the err
itself.  At that point we get the benefit of annotated errors in the
call stack and a non-nil error return so that callers don't carry on
thinking there was no error.

Fixes: #1068

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>