ciao-launcher: send networks in Ready frame
The Ready and Stats frames are almost symmetrical, but the latter has
additional details not needed by the ssntp.server / scheduler, but Ready
is kept smaller.  Scheduler deliberately does not parse Stats frames,
because they are potentially huge with detailed information on every
instance on a node.

To better enable placement of CNCI workloads, scheduler needs to know
about the networks available on a network node.  Which requires
promoting the networks list to additionally be sent in the Ready frame.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>