payloads: add PhysicalNetwork resource type for start yaml
A CNCI workload may need placed on a network node that is known to
have access to a particular subnet.  To do this we need an additional
resource type identifier.  This identifier will be associated with the
new optional resource string value instead of the resource quantity
integer value.

A subsequent patch to scheduler will use this field to make correct
placement decisions.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>