ciao-cli: Add functions to autogenerate template documentation
This commits adds some functions that can be used to auto-generate the
documentation that describes how templates may be used for ciao-cli
commands.  As, an example, to generate the template usage documentation
for a command that operates on slices of  CiaoCNCIs you might call

generateUsageUndecorated([]types.CiaoCNCI{})

and you would receive a string that looks something like this.

[]struct {
	ID        string `json:"id"`
	TenantID  string `json:"tenant_id"`
	IPv4      string `json:"IPv4"`
	Geography string `json:"geography"`
	Subnets   []struct {
		Subnet string `json:"subnet_cidr"`
	} `json:"subnets"`
}

There's another function, generateUsageDecorated that also includes
information about all the addtional functions ciao-cli provides for
use in templates.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>