openstack: Export extractProject functionality
Some service API requests in OpenStack will not contain the project/tenant
in them. It requires a mechanism for retrieving project data and make
the proper treatment of requests.

All API requests will contain a tokenID, this information can be used in
order to extract the tenant/project from identity service.

Fixes #1082

Signed-off-by: Obed N Munoz <obed.n.munoz@intel.com>