scheduler: add list of network nodes
Way way back in history Samuel Ortiz made a tweek to the scheduler's
compute nodes tracking in order to give a means of always seeing a
durably ordered list of compute nodes.  The list can make some access
patterns more efficient than the map.  And with just a map, the observed
node order can vary.  Because network nodes are typically fewer in number
than compute nodes, it wasn't such an issue there and was not implemented
to match.

This patch removes this inconsistency and adds a list to the network
node tracking.  This makes it easier for a follow on patch to implement
a deterministic unit test of the pickNetworkNode() function in a way
that is symmetric with the pickComputeNode() unit test.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>