ciao-down: Limit default CPU and memory resources
Previously, by default, ciao-down would allocate 50% of the host's
CPU and 50% of its available RAM to its VMs.  This can be a little
excessive on powerful machines that have many CPUs and lots of RAM.
Therefore, ciao-down now limits the defaults to 8 GB of RAM and 8 CPUs.
These settings can still be overridden by the user on the command line.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>