ciao-down: Delete vmStarted function
The vmStarted function was used to determine whether the Qemu instance was running
or not.  The problem is that this function relied on connecting to the instance's
QMP socket and only one process can connect to this socket at any one time.  This
meant that ./ciao-down connect just hung if you ran it while ./ciao-down prepare was
running.  Often this is a useful thing to do if you are trying to debug a failure
with the VM setup.  As vmStarted wasn't really adding very much and it didn't work
under all circumstances, it's best to just remove it.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>