travis: Allow go tip builds to fail
The go tip builds are currently failing due to a compiler error, i.e.,
a crash in the compiler.  This commit will allow ciao builds to continue
to pass even if go tip builds fail.  This is reasonable enough as there
are likely to be build breakages on go tip from time to time.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>