openstack: Omit UpdatedAt if empty
This commit modifies the tags for serializing image.DefaultResponse.
The UpdateAt field is optional and is currently not filled in by
our image service.  The latest version of gophercloud will refuse
to parse the json request if the UpdateAt field is null.  This commit
ensures that the serialized version of image.DefaultResponse does not
contain a null UpdateAt field avoiding the parse issues in gophercloud.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>