ciao-cli, bat, ciao-controller: Fix volume commands
Gophercloud has changed the way it handles TenantIDs in volume structures.
This information has been moved to a new structure.  This means we need
to modify ciao-cli and bat to retrieve this information.

It will also not parse the volume timestamps returned by ciao-controller.
This affects the CreatedAt and UpdatedAt timestamps returned as part of the
information about a volume.  We can hack around the problem for
ciao-cli volume show but not for ciao-cli volume list.  Therefore this
commit modifies ciao-controller to not send a CreatedAt timestamp when
processing the volumeListDetails command.  As far as I can tell from
the openstack documentation this field shouldn't be returned for this
endpoint anyway.  The openstack cli tool certainly ignores it.  One
consequence of all this is that it is no longer possible to sort the
results of ciao-cli volume list by creation time.  This commit sorts them
by name.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>