ciao-down: Allow qemu boot to continue even if 9p mounts fail
There seems to be a race condition that causes the mounting of the
shared folders via 9p to fail.  This in turn causes the ciao-down
VM boot to fail.  This commit switches to using the system.d
automount facility ( which may fix the mount problem ) and allows the
boot to continue even if the mounts fail.  This is preferable to the
current status in which ciao-down connect seems to hang from time
to time on some machines.  If automatic mounting fails it should
still be possible to manually mount the shared folders from the command
line.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>