ciao-controller: pull controller's FQDN from its certificate
Rather than having the user tell us in the ansible setup or when
creating certificates which host is their controller, when the
controller starts up and uses its certificate, it can find its
configured fully qualified domain name in the certificate.

If for some reason the certificate includes multiple names, the first is
used.  If none is given, an attempt is made to us the non-FQDN available
in os.Hostname().

This should give the intended hostname, without an additional human
touch point on configuration.

Fixes: #1138

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>