ciao-launcher, ciao-controller: Auto delete LaunchExit failures
Previously, instances that failed to launch but which were otherwise correctly
configured on the host node were not deleted by controller or launcher.  They
would appear in the instance list as exitted.  This behaviour no longer makes
sense given our new migration architecture.  Instances that are not running
no longer have any state on compute or network nodes.  For this reason we now
need to make sure that launcher and controller delete them.

Partial Fix for #1041

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>