ciao-launcher: Power down instances properly
Previously, launcher was quitting the qemu process instantly when stopping
and deleting instances.  This is no longer acceptable as instances deleted
from a compute node might be migrated elsewhere, so we need to ensure that
they are shutdown properly to avoid data loss.

We also need to update the DeleteAllInstances BAT test which waited 5
seconds for an instance to quit.  5 seconds is no longer enough so we
update the timeout to 15 seconds.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>