sqlite3db: Switch workload_resources table to use string resource name
The workload_resources table is used to the store the resource/value
combinations used by a workload; it is the database persistence of the
payloads.RequestedResource struct for a given workload.

Previously this was being stored as an integer referencing the resources
table which held a mapping back to the string representation.

The resources table has not been removed as it is used by the older
limits and usage datastore APIs which will be cleaned up in a later PR.

In order to ensure testability an ORDER BY was added to the SELECT
statement to ensure that the contents of the slice for the resources are
in a predictable order.

Fixes: #1114

Signed-off-by: Rob Bradford <robert.bradford@intel.com>