ciao-cli: Fix privilege related error reporting for quotas command
This change switches the error messages to use fatalf() and terminate
execution if there is a problem with the user privileges.

It also works to ensure that if the user is an admin then they must
specify a tenant id to check as the Ciao API url creation never embeds
the url for the the admin tenant and so the unprivileged path does not
work the admin user.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>