ssntp, payloads, launcher: Add InstanceStopped event
This commit adds a new SSNTP event InstanceStopped.  This event is sent by
launcher when an instance dies of its own accord or is deleted by an
SSNTP Delete command whose Migration flag is set to true.

This commit also adds the Migration flag to the payloads.StopCmd the
handling of this command in launcher and some extra unit tests to make
sure they correct event is sent when an instance is deleted, either
because it is being migrated or because it is being deleted permenantly
from the cluster.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>