ciao-controller: Update quotas with resources for instance lifecycle
When an instance that is not the CNCI is created then the resources used
are obtained from the workload defaults and marked as consumed in the
quota service.

If the instance start fails early (potentially caused by exceding the
quota) and Clean() is called then the resources will be released.

When scheduler fails to schedule an instance or if an instance is
deleted then the resources are released (again checking for CNCI so that
those resources that were never consumed are not released.)

Fixes: #1075

Signed-off-by: Rob Bradford <robert.bradford@intel.com>