ciao-controller: Delete ephemeral volumes on start failure
If the instance start fails due to e.g. the cloud being full then the
ephemeral volume needs to be deleted. This change makes controller
delete the ephemeral volume if it receives a start failure message from
scheduler if that error was a fatal error where the instance was never
started.

Tested by starting 60 instances on a small cluster and then observing
that without the patches the volumes for the rejected instances remain
after deleting all; with the patch no residual volumes remain.

Fixes: #1173

Signed-off-by: Rob Bradford <robert.bradford@intel.com>