systemd: Report to systemd when service has finished loading.
By setting Type=notify in systemd unit files, systemd will wait until
the service has actually been started to return success or fail if the
service does not start correctly.

The previous behaviour was that systemctl start ciao-*.service will
return successfully and immediately regardless of the state of the
service.

This change also eliminates some race-condition where automaed deployment
scripts continue to make use of the api when the service is not fully
started.

Fixes #1181

Signed-off-by: Alberto Murillo Silva <alberto.murillo.silva@intel.com>