ciao-controller: api: check for privilege
Modify our API Handler to include a boolean to indicate whether
a route should be used only be a privileged user or not. Modify
ServeHTTP to check the PrivKey context set by the identity service
to see if a user of a route has neccessary privileged. If not,
send http status Unauthorized.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>