ciao-controller: Implement instance migration
This commit adds support for instance migration in controller.  It does
this my mapping the stop command received from clients to an SSNTP
DELETE command and the restart command to an SSNTP start.  There are
ofcourse some complications.  A special flag, migration, is set in the
SSNTP DELETE command payload to indicate that the instance being deleted
is actually being migrated.  This causes launcher to send an InstanceStopped
event, rather than an InstanceDeleted event, when it has deleted the
instance from the local node.  Controller has been modified to remove
the association between a node and an instance and to set the instance
state to exited when it receives an InstanceStopped event, but to otherwise
leave the instance untouched.  Launcher also sends InstanceStopped events when
an instance dies of its own accord, e.g., its a container whose workload has
finished.

At a later stage when the user issues a restart command controller now
recreates the original START command from the information it stores about
the instance.  This command is sent to scheduler and the instance is
restarted on a new node.

Fixes #1121
Fixes #1043

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>