ciao-controller: configuration update API implementation
This commit handles the API request for a configuration operation
(update, show).

For the update case, it validates the input of the request, checking
that a sane value will replace current configuration element to be changed.

For the show case, a security cleanup is performed to avoid sensible
data (e.g: ssh-key, passwords) to be transfered/shown.

This commit fixes #1112

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>