ciao-controller: Report all successfully added instances
When adding multiple instances it can be possible for one to fail due to
exceeding a quota or some other error. When that happens stop starting
any more instances but report all those that did get added and generate
an event with the error.

If none got added then report the error as before.

Fixes: #1203

Signed-off-by: Rob Bradford <robert.bradford@intel.com>