ciao-controller: Ensure cache is correct after restart
There were two errors in datastore.Init that was causing controller to behave
incorrectly after a restart.  The instances stored in the ds.tenant structure
did not point to the instances in the ds.instances map.  This meant they did
not get updated when new stats were received.  Consequently, ciao-cli instance
list did not work after a restart.

The ds.instance.Attachments were not initialised after a restart of controller.
This meant that instances appeared to have no attachements.  This caused issues
for migration as it meant that exited instances could not be restarted after
controller had been rebooted.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>