ciao-controller: ciao-cli: do not use privileged route for instance list
ciao-cli is currently using a privileged route to support
"ciao-cli instance list -workload"

All calls to list instances should go through the non-privileged tenant
specific route. However, because normal users might still like to use
the -workload option, modify the non-privileged route to take a query
value of "flavor". The controller code to filter by workload UUID is
already supported, however, previously the "flavor" value was expected
to be passed via the URL. Modify the controller code to accept the
query parameter as well as the mux var for flavor.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>