ansible: Update documentation to run ansible on the deployment host
We used to provide a docker container with ansible and the dependencies
required by the playbooks already installed on it.

This commit updates the documentation to instruct the user to install
ansible and the required packages on the deployment host since using
the container proved to be more cumbersome than installing the few
depdendencies.

Signed-off-by: Alberto Murillo Silva <alberto.murillo.silva@intel.com>