ciao-launcher, controller, payloads:  Fix restart errors
This commit fixes two bugs related to restart.

1. Restarted instances that failed to restart could be deleted if
   launcher had send a STATS command to controller resetting the
   instance status to pending before the failure occurred.
2. Deleting instances that were restarting could lead to them
   becoming partially deleted.  Controller lost track of the
   instances but the instances themselves were still running.

The fix works by reseting the state of restarting instances to
pending, and added a flag to the start and start failure commands
to indicated whether an instance is being restarted or started.
Launcher uses the migration flag in the start payload to figure
to set the migration flag in the start failure payload.  Controller,
can use the presence of this command to decide whether it should
delete an instance that fails to start.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>