ssntp: implement configuration update storage in server
the new configuration payload recieved from the ciao-controller
is set in the SSNTP server configuiration and later saved by using
the configuration package.

A backup is defined when SSNTP server is started in order to enable
configuration backups previous to the configuration overwrite, allowing
the admin to preserve previous definitions of the cluster configuration.

This commit closes #1185

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>