ciao-controller: Fix workload's image ID reference in ciao-cli output
This commit fixes OpenStack's compatibility for Disk variable, it's not
the Image ID, it refers to the primary disk's size according to OpenStack
Flavor Details API method.
https://developer.openstack.org/api-ref/compute/?expanded=list-flavors-with-details-detail#list-flavors-with-details

A new CIAO specific API will be required in order to show image and storage
details. This should be addressed in a separate PR.

Addresses part of #1232

Signed-off-by: Obed N Munoz <obed.n.munoz@intel.com>